pyproject.toml
[project]
name = "vectome"
version = "0.0.1"
authors = [
  { name="Eachan Johnson", email="eachan.johnson@crick.ac.uk" },
]
description = "Deterministic vectorization of genomes."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = [
  "template",
]

classifiers = [  
  
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  #"Intended Audience :: Science/Research",
  #"Topic :: Scientific/Engineering :: Artificial Intelligence",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "bioino",
  "carabiner-tools>=0.0.4",
  "joblib",
  "numpy",
  "platformdirs>=4.4.0",
  "pyyaml",
  "requests",
  "sourmash",
  "tqdm",
]

[project.optional-dependencies]
all = []
#tf = ["duvida", "tensorflow"]
#jax = ["jax>=0.5.3"]

[project.urls]
"Homepage" = "https://github.com/scbirlab/vectome"
"Bug Tracker" = "https://github.com/scbirlab/vectome/issues"

[project.scripts] 
vectome = "vectome.cli:main"

[tool.setuptools]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {vectome = ["data/*.json"]}

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[test]
deps = ["pytest>=7.4", "pytest-cov"] 
